refactor(arch): Move generate_uuid() and get_secret_hash() to app.util
